+++ app/src/main/java/com/comicify/core/storage/DatabaseEntryPoint.kt
@@ -7,6 +7,7 @@ import dagger.hilt.components.SingletonComponent
 @EntryPoint
 @InstallIn(SingletonComponent::class)
 interface DatabaseEntryPoint {
+    fun comicDao(): ComicDao
     fun pageDetectionDao(): PageDetectionDao
     fun comicSettingsDao(): ComicSettingsDao
 }